Jotrick/refactor entities to common (#15)
* Moved repositories and entites to common project; Moved project specific data logic to extensions

* Added comment

* Added comment

* Added comment

* Proj file changes
