+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Common/Repositories/Notification/NotificationEntity.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // </copyright>
 
-namespace Microsoft.Teams.Apps.CompanyCommunicator.Repositories.Notification
+namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.Notification
 {
     using System.Collections.Generic;
     using Microsoft.Azure.Cosmos.Table;
