+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Common/Repositories/SentNotification/RecipientEntity.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // </copyright>
 
-namespace Microsoft.Teams.Apps.CompanyCommunicator.Repositories.SentNotification
+namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.SentNotification
 {
     /// <summary>
     /// Audience entity class.
