+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Common/Repositories/User/UserDataEntity.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // </copyright>
 
-namespace Microsoft.Teams.Apps.CompanyCommunicator.Repositories.User
+namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.User
 {
     using Microsoft.Azure.Cosmos.Table;
