+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.sln
@@ -1,9 +1,14 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29009.5
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.757
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Teams.Apps.CompanyCommunicator", "Microsoft.Teams.Apps.CompanyCommunicator\Microsoft.Teams.Apps.CompanyCommunicator.csproj", "{1EF4F8A5-2C38-4059-BD5C-E8478DE4779F}"
+	ProjectSection(ProjectDependencies) = postProject
+		{069D255F-37E5-49DA-90D2-E3942C64970C} = {069D255F-37E5-49DA-90D2-E3942C64970C}
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Teams.Apps.CompanyCommunicator.Common", "Microsoft.Teams.Apps.CompanyCommunicator.Common\Microsoft.Teams.Apps.CompanyCommunicator.Common.csproj", "{069D255F-37E5-49DA-90D2-E3942C64970C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +20,10 @@ Global
 		{1EF4F8A5-2C38-4059-BD5C-E8478DE4779F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1EF4F8A5-2C38-4059-BD5C-E8478DE4779F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1EF4F8A5-2C38-4059-BD5C-E8478DE4779F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{069D255F-37E5-49DA-90D2-E3942C64970C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{069D255F-37E5-49DA-90D2-E3942C64970C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{069D255F-37E5-49DA-90D2-E3942C64970C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{069D255F-37E5-49DA-90D2-E3942C64970C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
