+++ Source/Microsoft.Teams.Apps.CompanyCommunicator/Bot/TeamsDataCapture.cs
@@ -6,8 +6,9 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Bot
 {
     using System.Threading.Tasks;
     using Microsoft.Bot.Schema;
-    using Microsoft.Teams.Apps.CompanyCommunicator.Repositories.Team;
-    using Microsoft.Teams.Apps.CompanyCommunicator.Repositories.User;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.Team;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.User;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Repositories.Extensions;
 
     /// <summary>
     /// Service to capture teams data.
