+++ Source/Microsoft.Teams.Apps.CompanyCommunicator/Repositories/RepositoriesServiceCollectionExtensions.cs
@@ -5,11 +5,11 @@
 namespace Microsoft.Teams.Apps.CompanyCommunicator.Repositories
 {
     using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Teams.Apps.CompanyCommunicator.Repositories.ActiveNotification;
-    using Microsoft.Teams.Apps.CompanyCommunicator.Repositories.Notification;
-    using Microsoft.Teams.Apps.CompanyCommunicator.Repositories.SentNotification;
-    using Microsoft.Teams.Apps.CompanyCommunicator.Repositories.Team;
-    using Microsoft.Teams.Apps.CompanyCommunicator.Repositories.User;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.ActiveNotification;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.Notification;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.SentNotification;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.Team;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.User;
 
     /// <summary>
     /// Extension class for registering respository services in DI container.
